Added keymaps from my Lily58 configuration
Note: Currently these keymaps are not windows-compatible. In order to
make these windows-compatible, I either need to figure out how to make
macros (so I can dynamically detect if we're in Windows mode), or I'll
need to create a different window movement layer. The latter seems more
feasible with the current feature-set of QMK.